app/pyproject.toml
[tool.poetry]
name = "fess-text-vectorizer"
version = "0.1.0"
description = "Embedding Server for Fess"
authors = ["CodeLibs, Inc. <info@codelibs.co>"]

[tool.poetry.dependencies]
python = "^3.8"
ginza = "5.1.0"
ja-ginza-electra = "5.1.0"
transformers = "4.12.5"
spacy-transformers = "1.1.4"
ginza-transformers = "0.4.1"
fastapi = "0.74.1"
uvicorn = "0.17.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
